moved pages out of components folder
