+++ src/FE/FEsrc/components/App.tsx
@@ -4,16 +4,16 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { AuthProvider } from '../contexts/AuthContext';
 
 import MenuHeader from './MenuHeader';
-import LandingPage from './pages/LandingPage';
-import LoginPage from './pages/LoginPage';
-import RegisterPage from './pages/RegisterPage';
-import AboutUsPage from './pages/AboutUsPage';
-import ContactUsPage from './pages/ContactUsPage';
-import FAQPage from './pages/FAQPage'
-import TrialSearchPage from './pages/TrialSearchPage';
-import SaveTrialsPage from './pages/SavedTrialsPage';
-import ProfileCreationPage from './pages/ProfileCreation';
-import ListProfiles from './pages/ListProfiles';
+import LandingPage from '../pages/LandingPage';
+import LoginPage from '../pages/LoginPage';
+import RegisterPage from '../pages/RegisterPage';
+import AboutUsPage from '../pages/AboutUsPage';
+import ContactUsPage from '../pages/ContactUsPage';
+import FAQPage from '../pages/FAQPage'
+import TrialSearchPage from '../pages/TrialSearchPage';
+import SaveTrialsPage from '../pages/SavedTrialsPage';
+import ProfileCreationPage from '../pages/ProfileCreation';
+import ListProfiles from '../pages/ListProfiles';
 
 function App() {
